style: rename fqtn to type name for better code understanding #347
