+++ src/parser/resolver/accessors/c-sharp-collector.ts
@@ -1,0 +1,23 @@
+import { AbstractCollector } from "./abstract-collector.js";
+
+export class CSharpCollector extends AbstractCollector {
+    protected getAccessorsQuery(): string {
+        return `
+            (property_declaration
+                type: (_) @accessor_return_type
+                name: (identifier) @accessor_name
+            )
+            (field_declaration
+                (variable_declaration
+                    type: (_) @accessor_return_type
+                    .
+                    (variable_declarator (identifier) @accessor_name)
+                )
+            )
+            (method_declaration
+                type: (_) @accessor_return_type
+                name: (identifier) @accessor_name
+            )
+        `;
+    }
+}
