+++ src/parser/resolver/accessors/php-collector.ts
@@ -1,0 +1,7 @@
+import { AbstractCollector } from "./abstract-collector.js";
+
+export class PHPCollector extends AbstractCollector {
+    protected getAccessorsQuery(): string {
+        return "";
+    }
+}
