+++ src/parser/resolver/fqtn-collector.ts
@@ -1,13 +1,13 @@
 import { type ParsedFile } from "../metrics/metric.js";
 import { type Language } from "../../helper/language.js";
-import { type FQTNInfo } from "./fully-qualified-type-names/abstract-collector.js";
-import { Factory as NamespaceCollectorFactory } from "./fully-qualified-type-names/factory.js";
+import { type TypeInfo } from "./types/abstract-collector.js";
+import { Factory as NamespaceCollectorFactory } from "./types/factory.js";
 
 export class FqtnCollector {
     private readonly namespaceCollectorFactory = new NamespaceCollectorFactory();
-    private readonly languageToFilesMap = new Map<Language, Map<string, Map<string, FQTNInfo>>>();
+    private readonly languageToFilesMap = new Map<Language, Map<string, Map<string, TypeInfo>>>();
 
-    getFQTNsFromFile(parsedFile: ParsedFile): Map<string, FQTNInfo> {
+    getFQTNsFromFile(parsedFile: ParsedFile): Map<string, TypeInfo> {
         const { filePath, language } = parsedFile;
         if (this.languageToFilesMap.get(language) === undefined) {
             this.languageToFilesMap.set(parsedFile.language, new Map());
@@ -22,15 +22,15 @@ export class FqtnCollector {
         const collector = this.namespaceCollectorFactory.getCollector(parsedFile);
         const FQTNsMap =
             collector === undefined
-                ? new Map<string, FQTNInfo>()
-                : collector.getFQTNsFromFile(parsedFile);
+                ? new Map<string, TypeInfo>()
+                : collector.getTypesFromFile(parsedFile);
 
         this.languageToFilesMap.get(language)?.set(filePath, FQTNsMap);
 
         return FQTNsMap;
     }
 
-    getAllNamespaces(): IterableIterator<[Language, Map<string, Map<string, FQTNInfo>>]> {
+    getAllNamespaces(): IterableIterator<[Language, Map<string, Map<string, TypeInfo>>]> {
         return this.languageToFilesMap.entries();
     }
 }
