--- src/parser/resolver/fully-qualified-type-names/resolver-strategy/filename-resolver.ts
@@ -1,7 +1,0 @@
-import { type FQTNInfo } from "../abstract-collector.js";
-
-export class FilenameStrategy {
-    getFullyQTNs(): Map<string, FQTNInfo> {
-        return new Map();
-    }
-}
