+++ src/parser/resolver/public-accessor-collector.ts
@@ -1,14 +1,14 @@
 import { type ParsedFile } from "../metrics/metric.js";
-import { Factory as AccessorCollectorFactory } from "./call-expressions/factory.js";
-import { type Accessor } from "./call-expressions/abstract-collector.js";
-import { type FQTNInfo } from "./fully-qualified-type-names/abstract-collector.js";
+import { Factory as AccessorCollectorFactory } from "./accessors/factory.js";
+import { type Accessor } from "./accessors/abstract-collector.js";
+import { type TypeInfo } from "./types/abstract-collector.js";
 
 export class PublicAccessorCollector {
     private readonly accessorCollectorFactory = new AccessorCollectorFactory();
 
     getPublicAccessorsFromFile(
         parsedFile: ParsedFile,
-        FQTNsFromFile: Map<string, FQTNInfo>,
+        FQTNsFromFile: Map<string, TypeInfo>,
     ): Map<string, Accessor[]> {
         const collector = this.accessorCollectorFactory.getCollector(parsedFile);
         const accessors = collector?.getPublicAccessorsFromFile(parsedFile, FQTNsFromFile);
