+++ src/parser/resolver/types/resolver-strategy/filename-resolver.ts
@@ -1,0 +1,7 @@
+import { type TypeInfo } from "../abstract-collector.js";
+
+export class FileNameStrategy {
+    getTypes(): Map<string, TypeInfo> {
+        return new Map();
+    }
+}
