+++ src/parser/resolver/types/resolver-strategy/types-query-strategy.ts
@@ -1,6 +1,6 @@
 import { debuglog, type DebugLoggerFunction } from "node:util";
 import { type QueryCapture } from "tree-sitter";
-import { type FQTNInfo } from "../abstract-collector.js";
+import { type TypeInfo } from "../abstract-collector.js";
 import { QueryBuilder } from "../../../queries/query-builder.js";
 import { formatCaptures } from "../../../../helper/helper.js";
 import { type ParsedFile } from "../../../metrics/metric.js";
@@ -9,26 +9,27 @@ import { SimpleQueryStatement } from "../../../queries/query-statements.js";
 let dlog: DebugLoggerFunction = debuglog("metric-gardener", (logger) => {
     dlog = logger;
 });
+type FileName = string;
+type TypeName = string;
+export class TypesQueryStrategy {
+    protected fileToTypesMap = new Map<FileName, Map<TypeName, TypeInfo>>();
 
-export class QueryStrategy {
-    protected fileToFQTNsMap = new Map<string, Map<string, FQTNInfo>>();
-
-    getFQTNsFromFile(
+    getTypesFromFile(
         parsedFile: ParsedFile,
         namespaceDelimiter: string,
-        namespacesQuery: string,
-    ): Map<string, FQTNInfo> {
+        typesQuery: string,
+    ): Map<string, TypeInfo> {
         const { filePath, language, tree } = parsedFile;
 
-        let FQTNsMap = this.fileToFQTNsMap.get(filePath);
-        if (FQTNsMap !== undefined) {
-            return FQTNsMap;
+        let typesMap = this.fileToTypesMap.get(filePath);
+        if (typesMap !== undefined) {
+            return typesMap;
         }
 
-        FQTNsMap = new Map<string, FQTNInfo>();
+        typesMap = new Map<TypeName, TypeInfo>();
 
         const queryBuilder = new QueryBuilder(language);
-        queryBuilder.setStatements([new SimpleQueryStatement(namespacesQuery)]);
+        queryBuilder.setStatements([new SimpleQueryStatement(typesQuery)]);
 
         const query = queryBuilder.build();
         let captures: QueryCapture[] = [];
@@ -38,7 +39,7 @@ export class QueryStrategy {
 
         const textCaptures = formatCaptures(tree, captures);
 
-        dlog("namespace definitions", filePath, textCaptures);
+        dlog("types definitions", filePath, textCaptures);
 
         for (let index = 0; index < textCaptures.length; index += 1) {
             const namespaceName = textCaptures[index].text;
@@ -57,16 +58,16 @@ export class QueryStrategy {
 
             while (hasClassDeclaration) {
                 const className = textCaptures[index].text;
-                const namespaceDeclaration: FQTNInfo = {
+                const typeDeclaration: TypeInfo = {
                     namespace: namespaceName,
-                    className,
+                    typeName: className,
                     classType: isInterface ? "interface" : "class",
-                    source: parsedFile.filePath,
+                    sourceFile: parsedFile.filePath,
                     namespaceDelimiter,
-                    implementedClasses: [],
+                    implementedFrom: [],
                 };
 
-                FQTNsMap.set(namespaceName + namespaceDelimiter + className, namespaceDeclaration);
+                typesMap.set(namespaceName + namespaceDelimiter + className, typeDeclaration);
 
                 // Jump to any next capture
                 index++;
@@ -75,7 +76,7 @@ export class QueryStrategy {
                 }
 
                 if (textCaptures[index]?.name === "extended_class") {
-                    namespaceDeclaration.extendedClass = textCaptures[index].text;
+                    typeDeclaration.extendedFrom = textCaptures[index].text;
                     index++;
                 }
 
@@ -87,8 +88,8 @@ export class QueryStrategy {
 
                 while (hasImplements) {
                     const implementedClassName = textCaptures[index].text;
-                    if (!namespaceDeclaration.implementedClasses.includes(implementedClassName)) {
-                        namespaceDeclaration.implementedClasses.push(implementedClassName);
+                    if (!typeDeclaration.implementedFrom.includes(implementedClassName)) {
+                        typeDeclaration.implementedFrom.push(implementedClassName);
                     }
 
                     // Jump to next capture, no matter if implements class or any other
@@ -116,8 +117,8 @@ export class QueryStrategy {
             }
         }
 
-        this.fileToFQTNsMap.set(filePath, FQTNsMap);
+        this.fileToTypesMap.set(filePath, typesMap);
 
-        return FQTNsMap;
+        return typesMap;
     }
 }
