+++ src/parser/resolver/usages-collector.ts
@@ -1,10 +1,10 @@
 import { type ParsedFile } from "../metrics/metric.js";
-import { Factory as UsageCollectorFactory } from "./type-usages/factory.js";
+import { Factory as UsageCollectorFactory } from "./call-expressions/factory.js";
 import { type FqtnCollector } from "./fqtn-collector.js";
 import {
     type TypeUsageCandidate,
     type UnresolvedCallExpression,
-} from "./type-usages/abstract-collector.js";
+} from "./call-expressions/abstract-collector.js";
 
 export class UsagesCollector {
     private readonly usageCollectorFactory = new UsageCollectorFactory();
